053: Extract mixins for Image/Loading & Saving, Image/Pixels, Shape/Attributes, Transform.
